pyproject.toml
[project]
name = "auto-pdfocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ocrmypdf>=16.13.0",
    "ulid-py>=1.1.0",
  ]
